refactor: log codebase
